# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FinSense AI API",
  "description": "Financial analytics, portfolio optimisation, ML forecasting and RAG research API. Decision support only - not investment advice.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/admin/imports/{kind}": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Import Csv",
    "operationId": "import_csv_api_v1_admin_imports__kind__post",
    "parameters": [
     {
      "name": "kind",
      "in": "path",
      "required": true,
      "schema": {
       "enum": [
        "benchmarks",
        "benchmark_values",
        "funds",
        "nav",
        "aum",
        "sip_flows",
        "holdings"
       ],
       "type": "string",
       "title": "Kind"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_csv_api_v1_admin_imports__kind__post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Import Csv Api V1 Admin Imports  Kind  Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/imports": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "List Imports",
    "operationId": "list_imports_api_v1_admin_imports_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 25,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Imports Api V1 Admin Imports Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/imports/{batch_id}": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Get Import",
    "operationId": "get_import_api_v1_admin_imports__batch_id__get",
    "parameters": [
     {
      "name": "batch_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Batch Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Import Api V1 Admin Imports  Batch Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/rag/diagnostics": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Rag Diagnostics",
    "description": "Full retrieval trace for the administrator's own document access.\n\nIncludes scores from every stage and the exact prompt that would be sent\nto the language model. It runs with the admin's permissions, so it never\nexposes other users' private documents.",
    "operationId": "rag_diagnostics_api_v1_admin_rag_diagnostics_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DiagnosticsRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Rag Diagnostics Api V1 Admin Rag Diagnostics Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/audit": {
   "get": {
    "tags": [
     "admin"
    ],
    "summary": "Audit Log",
    "operationId": "audit_log_api_v1_admin_audit_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Audit Log Api V1 Admin Audit Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/admin/users/{user_id}": {
   "patch": {
    "tags": [
     "admin"
    ],
    "summary": "Update User",
    "operationId": "update_user_api_v1_admin_users__user_id__patch",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Update User Api V1 Admin Users  User Id  Patch"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/admin/amfi/register": {
   "post": {
    "tags": [
     "admin"
    ],
    "summary": "Amfi Register",
    "operationId": "amfi_register_api_v1_admin_amfi_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AmfiRegister"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Amfi Register Api V1 Admin Amfi Register Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AmfiRegister": {
    "properties": {
     "scheme_code": {
      "type": "string",
      "pattern": "^\\d{3,8}$",
      "title": "Scheme Code"
     },
     "category": {
      "type": "string",
      "maxLength": 80,
      "minLength": 2,
      "title": "Category"
     },
     "asset_class": {
      "type": "string",
      "enum": [
       "equity",
       "debt",
       "hybrid",
       "other"
      ],
      "title": "Asset Class"
     }
    },
    "type": "object",
    "required": [
     "scheme_code",
     "category",
     "asset_class"
    ],
    "title": "AmfiRegister"
   },
...
   "Body_import_csv_api_v1_admin_imports__kind__post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     },
     "source_code": {
      "type": "string",
      "enum": [
       "csv-import",
       "synthetic-demo"
      ],
      "title": "Source Code",
      "default": "csv-import"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_import_csv_api_v1_admin_imports__kind__post"
   },
...
   "DiagnosticsRequest": {
    "properties": {
     "query": {
      "type": "string",
      "maxLength": 1000,
      "minLength": 2,
      "title": "Query"
     },
     "previous_questions": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 5,
      "title": "Previous Questions"
     },
     "mode": {
      "type": "string",
      "enum": [
       "hybrid",
       "vector",
       "lexical"
      ],
      "title": "Mode",
      "default": "hybrid"
     }
    },
    "type": "object",
    "required": [
     "query"
    ],
    "title": "DiagnosticsRequest"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "UserUpdate": {
    "properties": {
     "role": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "user",
         "admin"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Role"
     },
     "is_active": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Is Active"
     }
    },
    "type": "object",
    "title": "UserUpdate"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}